presentation/requirements.txt
asttokens==2.4.1
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
executing==2.1.0
Flask==3.0.3
fonttools==4.54.1
h11==0.14.0
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
robotframework==7.1.1
robotframework-pythonlibcore==4.4.1
robotframework-seleniumlibrary==6.6.1
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
xlrd==2.0.1
